+++ src/extension.ts
@@ -5,9 +5,10 @@ import * as Commands from './commands';
 import * as Utility from './utility/index';
 
 export async function activate(context: vscode.ExtensionContext) {
-    Utility.state.init(context);
+    Utility.context.init(context);
 
     // !TODO - Refactor command registration, there's a common thread here.
+    // I really need to refactor this, this is cursed.
     if (!Commands.shared.installed.installHeaders) {
         Commands.installHeaders.listen(context);
     }
@@ -39,6 +40,22 @@ export async function activate(context: vscode.ExtensionContext) {
     if (!Commands.shared.installed.addKeyToWallet) {
         Commands.addKeyToWallet.register(context);
     }
+
+    if (!Commands.shared.installed.lockWallet) {
+        Commands.lockWallet.register(context);
+    }
+
+    if (!Commands.shared.installed.unlockWallet) {
+        Commands.unlockWallet.register(context);
+    }
+
+    if (!Commands.shared.installed.transact) {
+        Commands.transact.register(context);
+    }
+
+    if (!Commands.shared.installed.deployContract) {
+        Commands.deployContract.register(context);
+    }
 }
 
 export function deactivate() {
@@ -50,4 +67,7 @@ export function deactivate() {
     Commands.showWalletPublicKeys.dispose();
     Commands.destroyWallet.dispose();
     Commands.addKeyToWallet.dispose();
+    Commands.lockWallet.dispose();
+    Commands.unlockWallet.dispose();
+    Commands.deployContract.dispose();
 }