+++ src/service/index.ts
@@ -1,1 +1,2 @@
 export * as wallet from './wallet';
+export * as api from './api'